020-latest-layer-version-macro/latest_layer_version_macro/requirements.txt
logmatic-python==0.1.7
